move code to StockList container and clean App component
